requirements.txt
brotli==1.1.0
ConfigParser==6.0.0
cryptography==41.0.5
Cython==3.0.5
dl==0.1.0
docutils==0.20.1
google-cloud-aiplatform==1.36.4
HTMLParser==0.0.2
importlib_metadata==6.8.0
keyring==24.3.0
langchain==0.0.339
lockfile==0.12.2
lxml==4.9.3
matplotlib==3.7.4
numpy==1.24.4
ordereddict==1.1
pandas==2.0.3
Pillow==10.1.0
protobuf==4.25.1
pydantic==1.10.8
pyOpenSSL==23.3.0
streamlit==1.28.2
streamlit_gchart==0.1.8
#thread==0.1.2
toml==0.10.2
tornado==6.3.3
#xmlrpclib==1.0.1
zipp==3.17.0
